allow test on other db
