+++ database/database_test.go
@@ -1,8 +1,6 @@
 package database
 
 import (
-	"errors"
-	"os"
 	"testing"
 
 	"github.com/gotify/server/test"
@@ -23,7 +21,13 @@ type DatabaseSuite struct {
 
 func (s *DatabaseSuite) BeforeTest(suiteName, testName string) {
 	s.tmpDir = test.NewTmpDir("gotify_databasesuite")
-	db, err := New("sqlite3", s.tmpDir.Path("testdb.db"), "defaultUser", "defaultPass", 5, true)
+	db, err := New(
+		test.GetEnv("TEST_DB_DIALECT", "sqlite3"),
+		test.GetEnv("TEST_DB_CONNECTION", s.tmpDir.Path("testdb.db")),
+		"defaultUser",
+		"defaultPass",
+		5,
+		true)
 	assert.Nil(s.T(), err)
 	s.db = db
 }
@@ -38,35 +42,4 @@ func TestInvalidDialect(t *testing.T) {
 	defer tmpDir.Clean()
 	_, err := New("asdf", tmpDir.Path("testdb.db"), "defaultUser", "defaultPass", 5, true)
 	assert.Error(t, err)
-}
-
-func TestCreateSqliteFolder(t *testing.T) {
-	tmpDir := test.NewTmpDir("gotify_testcreatesqlitefolder")
-	defer tmpDir.Clean()
-
-	db, err := New("sqlite3", tmpDir.Path("somepath/testdb.db"), "defaultUser", "defaultPass", 5, true)
-	assert.Nil(t, err)
-	assert.DirExists(t, tmpDir.Path("somepath"))
-	db.Close()
-}
-
-func TestWithAlreadyExistingSqliteFolder(t *testing.T) {
-	tmpDir := test.NewTmpDir("gotify_testwithexistingfolder")
-	defer tmpDir.Clean()
-
-	db, err := New("sqlite3", tmpDir.Path("somepath/testdb.db"), "defaultUser", "defaultPass", 5, true)
-	assert.Nil(t, err)
-	assert.DirExists(t, tmpDir.Path("somepath"))
-	db.Close()
-}
-
-func TestPanicsOnMkdirError(t *testing.T) {
-	tmpDir := test.NewTmpDir("gotify_testpanicsonmkdirerror")
-	defer tmpDir.Clean()
-	mkdirAll = func(path string, perm os.FileMode) error {
-		return errors.New("ERROR")
-	}
-	assert.Panics(t, func() {
-		New("sqlite3", tmpDir.Path("somepath/test.db"), "defaultUser", "defaultPass", 5, true)
-	})
 }
